duplicate `CleanUp` button to ExerciseNode - add base case of todomvc to its chapter - hide `ShowExerciseLog` button
